pyproject.toml
[project]
name = "centr-invest-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "dishka>=1.7.2",
    "fastapi[standard]>=0.123.10",
    "joblib==1.5.2",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn>=0.38.0",
    "category-classifier>=0.1.0",
    "python-multipart>=0.0.20",
]

[tool.uv.workspace]
members = [
    "third-party/*"
]

[tool.uv.sources]
category-classifier = { workspace = true }

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "ruff>=0.14.8",
]

[tool.mypy]
exclude = [
    "third-party/*",
    "app/alembic",
]

[tool.ruff]
exclude = [
    "third-party/*",
    "app/alembic",
]
